rooms-list - order by buttons working
When you click on the order by buttons it posts
a form, but that form data doesnt contain any of
the filters.  Which means that the filters get
lost when you re-order.

To get around this I used session variables to
store roomListOrder and roomListFilters. Now when
ordering req.session.roomListFilters can be used
to re-apply the filters as well as the order.

I implemented the ordering by creating an
"ORDER BY" string based on which button was
pressed on the form, and passing this to the
getRooms function along with the filters. This
"ORDER BY" string is appended to the SQL
statement. When the filter form is submitted
it uses req.session.roomListOrder to work out
what the current order of the list is.

The original code used app.post("rooms-list") as
a target when the form was submitted. There is
already an app.get("rooms-list"). To simplify
this I removed the app.post("rooms-list") and put
the ordering and the filtering code in
app.post("rooms-list-filtered").  Ordering and
filtering occur if the order by button
is pressed OR the filter is applied.