Week 6 Security: Jquery, CORS & CSP.
1. Jquery updated to latest version (3.7.1)

2. Cross Origin Resource Sharing

This was highlighted in the security report,
because we are now using ajax.
New headers were introduced to fix this issue:

  Access-Control-Allow-Origin:
    http://localhost:8000
  // restrict the types of access methods we allow
  // only need get and post, not options, put or
  //delete
  Access-Control-Allow-Methods: GET, POST
  Access-Control-Allow-Headers: Content-Type,
    Authorization,
    X-Requested-With,
    If-Modified-Since

3. Content Security Policy Updates

Ben reported issues with in-line scripts before so
the policy was commented.  in-line scripts are now
allowed - but this does flag a meduim security
issue.

The new Policy is below.

*Note this means that any resources not specified
here will NOT work.*

For example using external images not specified
in the domains below will mean the image won't
load. The server name will need to be changed
when the app is installed - I'll look into making
this a variable in the .env to make installation
simpler.
  default-src http://localhost:8000;
  // unsafe-inline allows inline scripts
  script-src http://localhost:8000 'unsafe-inline'
    https://ajax.googleapis.com;
  img-src 'self' https://cdn3.iconfinder.com
    https://i.imgur.com;
  style-src http://localhost:8000;
  font-src http://localhost:8000;
  // stops the page being used within
  // external websites
  frame-ancestors http://localhost:8000;
  // only allow form submission from our pages
  form-action http://localhost:8000;