add-booking.ejs working routes: add-booking, add-booking-submit functions: insertUpdateBooking,   insertUpdateRiskAssessment how to use 1. go to the book room menu item 2. filter on a date AND a timeslot 3. when you have done that the page will show rooms that do not overlap with the date and timeslot you have chosen. 4. There will now be a book button with the date and time shown that you will be booking for.  To change this date and time - change the filters and confirm them and the button will change. 5. click on the button and you will go to add-booking. 7. SQL to test that it's worked select max(id) from booking; which tells you the last booking select * from booking where id = *the number you got from before* same with risk assessment.
issues
1. the buttons don't display properly
at the bottom of the form.  I couldn't
get this to work without changing the
styling for the other pages that use
that styling - hopefully someone can
find an easy fix for that.
2. when you add the booking, it just
shows a confirmation.  In the final
version I guess we just go back to the
menu or the rooms-list.
3. there are lots of things here
that could be done in v1.  You can
book a room for a date less than today,
for example and many other things.

other comments.
in app.get(/add-booking) there is
commented suggested SQL for getting
the booking and risk_assessment data
back. I didn't need to do that in
add-booking, but that would be needed
in view-booking and edit-booking for
example

in insertUpdateBooking I added some
examples for doing updates.  The mode
parameter just changes what SQL is run
so more modes can be added if needed.