PART C done (ISH)
Functional view-booking page works.
Needed to add relative paths to the logos otherwise css complained.
Dunno why.

getbookingsbyid function used to do the obvious this (its in the name)

bookings-list button in the html now references a request page that takes the parameters of the booking id and is used for the query which takes you to the booking page.

See details of the push for more details. Im fried