Part D. completed edit-room
Created a list of rooms called edit-room-list.ejs
this allows the user to select the room using
the filters.
There is an edit button that sends the ID to
the edit-room page.
This is different to the hi-fidelity plan, so
if everyone considers this a bad solution then
this can be changed in v1. The existing code
passed all of the room data to the edit-room page
and created a select with the room number in.
To get that working a javascript event would need
to be coded in the page to fill in all of the
fields when the select button was used to choose
a room, including a hidden field that contained
the room.id.  For speed it was easier to re-use
the rooms-list code. A pro of this is that the
user can find the room to edit more easily.

1. copied rooms-list.ejs to edit-rooms-list.ejs
2. removed the code in the ejs that only showed
the button when a date and timeslot was selected
in the filter. Changed button to "Edit" and
posting to edit-room
3. altered the .ejs title and heading
4. copied rooms-list & rooms-list-filtered
routes to edit-rooms-list and
edit-rooms-list-filtered
5. edited edit-rooms-list and
edit-rooms-list-filtered replacing room with
editRoom
6. no need to change getRooms function
7. update admin menu in login-success to
add Edit Rooms
8. changed app.get("/edit-room") to
app.post("/edit-room")
9. edited edit-room to display the data
passed by the sql in the fields.
10. edit-booking working
11. converted edit-booking to use
Promise.all so could get multiple data
sets one for roomTypes, one for rooms
and one for building names.  this is
so valid select option can be shown
on the edit room page, had to pass
room id and the page name as new
params
12. edit-room select inputs: changed
so the current value from the database
was the selected one, for roomType,
buildingName and isAcceptingBookings.
12. on success returned to login-success
it's possible to return to edit-rooms-list
but needs another promise.all.  Consider
for v1
13.  use room 1000 for testing so the real
data doesn't get damaged