(#1 complete) added a counter for 200 characters when inputting risk assessments.
pretty simple - used a couple scripts to make sure the text can't go over 200 characters for edit-booking and add-booking.

Didn't do it for review-booking as those are read only, but it's the same code if we need to.

Also fixed a minor issue with add-booking where you could still add dates from the past if you typed them in manually. these changes mean you can no longer do that (it resets immediately to today's date if the user tries).

yahoo.