(#18 complete) added a reset filters button
Had to reshuffle a couple things so that everything works as expected - i've defined the updateListData function explicitly now; tiny bit hacky but it works.

Had to use a tiny delay when resetting, otherwise it would handle the data before resetting the form and the list would not change back to the default.

Also made the button red to show that resetting is a drastic change; can be changed later if it doesnt fit our aesthetic.