(#31 complete) - all list images are now the same size; they shrink and grow with the window
images are now all the same size and shape and they shrink and grow with the window.

Some minor things were changes, affecting responsiveness. Widths were changed to 100% on all list pages meaning text collapses rather than disappears with the page (could be better than before? Down to personal pref.)

Buttons for ordering-by are now different widths on each page. this can be changed later if need be; the functionality is all there.