2FA V1 is Complete  Activated
I have used speakeasy to implement 2fa for the application.

You need to install google authenticator and scan the code and input the returned code into the input box after logging in.

There is a boolean variable activateTwoFactor that activates 2fa for the entire application.

If you have issues, deactivate this by setting it to false and log in as normal.