#16 mostly completed
All of the lists page has their filter buttons onto the header allowing the user to be able to see more rooms on the page. However this has required me to create multiple header files for each specific list page other than add/edit room as they share the same buttons. Will require further research on how to do this with only one file to change the headers based on which list page.