issue #46 added switches to .env to allow for testing database hosting at uni but also keeping cloud database option.
This is so we can keep running with planetscale
for dev but I can still test hosting at uni.

When I have finished testing I will publish
the additional .env variables.  No need
for anyone to change anything at the moment.

Also added a TWO_FACTOR=true or false variable
to .env.  Setting this in .env overrides
the activateTwoFactor variable in app.js
if that TWO_FACTOR= is not in .env
activateTwoFactor in app.js is used.