Update app.js
Issue #46 CORS security added back in.  New variable put in .env  (PRODUCTION_URL=) if this is not present the code will use localhost:8000 which means dev will still work. If there is a PRODUCTION_URL in .env it will be used for the CORS security headers.  I can now test this part of the security on the uni server and not break dev.